update (#1)
* airlock sync (#1)

* airlock sync

* empty commit

* Update docker-compose.plaintext.proxy.yml

* Update docker-compose.plaintext.yml

* Update docker-compose.plaintext.backup-and-restore.yml

* Update docker-compose.plaintext.proxy.backup-and-restore.yml

* Update connect-onprem-to-cloud.sh

* Update connect-onprem-to-cloud-avro.sh

* Update connect-onprem-to-cloud-avro-with-sr-basic-auth.sh

* Update cli_function.sh

* Update get-number-records.sh

* Update heap-dump.sh

* Update snippets.sh

* Update create-or-update.sh

* Update bashly.yml

* Features3 (#4)

* reproduction-models

* fix self certs

* 2.00.076.00.20240701.1

* wip (#5)

* update

* bump to 3.24.4

* fix for jdk17

* fix schema

* cleanup pushtopic

* Add HTTP Sink V2 Fully Managed example #5885

* wip (#6)

* minor

* minor

* delete topic

* Add HTTP Source V2 Fully Managed example #5884

* sleep 60

* fix

* wip (#7)

* debug

* MyLeadPushTopicsV2

* wip

* 7.7.1

* playground connector show-lag

* wip (#8)

* doing auth again

* 3.25.1

* wip

* wip

* jdk 17 fix

* 16092024 2 (#9)

* Add example for jsonata #5756

* check does not work with CI for some reasons, only relying on lag

* Issue with adding tcpdump #5917

* wip (#10)

* separate name for cloud

* delete

* Add Google Cloud Functions Gen 2 Sink Connector #5869

* update (#12)

* 26092024 2 (#13)

* Add --initial-state option to "playground connector create-or-update" #5929

* faker location

* wip (#14)

* 3.5.3

* playground connector show-lag issue when used with --environment ccloud example #5943

* remove connect-jms-sag-um-source and connect-jms-sag-um-sink examples as Docker image is no longer provided by SAG #5949

* wip

* wip (#15)

* https://github.com/vdesabou/kafka-docker-playground/issues/5187

* delete retry

* test

* wip

* Add custom connector example for Tabular Iceberg sink #5971

* minor

* topic produce: make nb_max_messages_to_generate configurable #5976

* fix on linux

+++ scripts/cli/src/commands/stop.sh
@@ -5,11 +5,17 @@ then
     logerror "File $test_file retrieved from $root_folder/playground.ini does not exist!"
     exit 1
 fi
+export flink_connectors=""
 filename=$(basename -- "$test_file")
 test_file_directory="$(dirname "${test_file}")"
 
 tmp_dir=$(mktemp -d -t pg-XXXXXXXXXX)
-trap 'rm -rf $tmp_dir' EXIT
+if [ -z "$PG_VERBOSE_MODE" ]
+then
+    trap 'rm -rf $tmp_dir' EXIT
+else
+    log "🐛📂 not deleting tmp dir $tmp_dir"
+fi
 
 log "🛑 Stopping example $filename in dir $test_file_directory"
 docker_command=$(playground state get run.docker_command)
